refactor + fix bug with OHMS
