+++ src/index.js
@@ -1,18 +1,9 @@
 import SlateTranscriptEditor from './components/index.js';
 import { secondsToTimecode, timecodeToSeconds, shortTimecode } from './util/timecode-converter/index.js';
 import convertDpeToSlate from './util/dpe-to-slate/index.js';
-import converSlateToDpe, { convertSlateToDpeAsync } from './util/export-adapters/slate-to-dpe/index.js';
+import converSlateToDpe from './util/export-adapters/slate-to-dpe/index.js';
 import slateToText from './util/export-adapters/txt';
 
 export default SlateTranscriptEditor;
 
-export {
-  SlateTranscriptEditor,
-  secondsToTimecode,
-  timecodeToSeconds,
-  shortTimecode,
-  convertDpeToSlate,
-  converSlateToDpe,
-  convertSlateToDpeAsync,
-  slateToText,
-};
+export { SlateTranscriptEditor, secondsToTimecode, timecodeToSeconds, shortTimecode, convertDpeToSlate, converSlateToDpe, slateToText };
